per conversation with Troy, trashing cobra init content and doing more from scrach using docs. have put command error working, but not put command working debugging does some funky stuff right now
