+++ README.md
@@ -41,6 +41,8 @@ https://github.com/spf13/cobra/pull/817
 
 - Make sub commands shared to convert to functions: https://stackoverflow.com/questions/43747075/cobra-commander-how-to-call-a-command-from-another-command
 
+- Cobra Documentation: https://github.com/spf13/cobra
+
 #### Practice CLI project
 - See https://github.com/FavoredFortune/CobraCLI for application with instructions in the [README](https://github.com/FavoredFortune/CobraCLI/blob/master/README.md)
